requirements.txt
flask
torch
pandas
Pillow
open_clip_torch